+++ CHANGES.md
@@ -15,3 +15,4 @@
 - **Infinite Scrolling:** UI change, random photo added
 - **Typing game:** Added fetching random word, added saving scores to scoreboard
 - **Speech Reader:** Added custom user's card with image select from Unsplash API, saved in LocalStorage
+- **Memory Cards:** UI change, added light and dark mode